array cancellation

+++ codeforces/codeforces round #668 Div 2/2 B array cancellation.cpp
@@ -98,7 +98,7 @@ int main(){
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);  
 
-    int t, n;   
+    ll t, n;      
     cin >> t;   
     while(t--){
         cin >> n;
